+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 | Version | Date       | Brief Description      | Authors        |
 | ------ | ---------- | ----------------------------- | ------------- |
+| 1.1.9 | 2023/12/13 | Improved mobile compatibility for the ticker. Added crypto, currency, and index information to the homepage. Changed display of country stocks. Removed navbar return button for mobile users. Enhanced account creation spam prevention. Reduced password complexity. | @behindsecurity |
 | 1.1.8 | 2023/12/05 | Added login system. Fixed news ticker. No other changes or security updates. | @behindsecurity |
 | 1.1.7 | 2023/12/04 | Added nation statistics, click removal for inactive news, and main stock display. Removed unnecessary scripts. Fixed error page and UTC-related bug. Improved captcha security. No other changes. | @behindsecurity |
 | 1.1.6 | 2023/11/24 | Implemented sleek accordion design with mobile detection, enriched news item details on comments page, streamlined navbar, and improved CLS. Changed source color, added placeholder text, revamped logo font, and relocated search bar. Removed outdated images, and fixed card height for mobile users. | @behindsecurity |
@@ -16,6 +17,25 @@
 | 1.0.1 | 2023/10/28 | Basically improved performance and made the code more modularized. | @behindsecurity |
 
 
+## 1.1.9 2023/12/13
+
+### Enhancements (2)
+- IME01: Enhanced mobile compatibility (ticker).
+- IME02: Crypto, countries currency and indexes information are now collected and displayed in the home page. 
+
+### Changed (1)
+- IMC01: Changed the way country stocks are displayed.
+
+### Removed (1)
+- IMR01: Removed navbar 'Return' button for mobile users.
+
+### Fixed (0)
+
+### Security (2)
+- IMS01: Implemented measures to prevent spam upon account creation.
+- IMS02: Decreased password policy complexity (current: 8 characters minimum, with at least 1 number).
+
+
 ## 1.1.8: 2023/12/05
 
 ### Enhancements (1)
@@ -26,7 +46,7 @@
 ### Removed (0)
 
 ### Fixed (1)
-- IMF01: Fixed news ticker
+- IMF01: Fixed news ticker.
 
 ### Security (0)